🚚 Move Networking test extensions to Networking module
