+++ Components/Sources/Networking/Testing/URL+TestData.swift
@@ -1,6 +1,8 @@
+#if DEBUG
 import Foundation
 
 extension URL {
     static var ackeeCZ = URL(string: "https://ackee.cz")!
     static var ackeeDE = URL(string: "https://ackee.de")!
 }
+#endif
