+++ Components/Sources/Networking/Testing/URLRequest+TestData.swift
@@ -1,3 +1,4 @@
+#if DEBUG
 import Foundation
 
 extension URLRequest {
@@ -10,3 +11,4 @@ extension URLRequest {
         return request
     }
 }
+#endif
